+++ packages/twenty-server/src/engine/core-modules/file/file-upload/resolvers/file-upload.resolver.ts
@@ -9,6 +9,7 @@ import { getWorkspaceAuthContext } from 'src/engine/core-modules/auth/storage/wo
 import { FileWithSignedUrlDTO } from 'src/engine/core-modules/file/dtos/file-with-sign-url.dto';
 import { FileUploadTargetDTO } from 'src/engine/core-modules/file/file-upload/dtos/file-upload-target.dto';
 import { FileUploadGraphqlApiExceptionFilter } from 'src/engine/core-modules/file/file-upload/filters/file-upload-graphql-api-exception.filter';
+import { CreateFileUploadPermissionGuard } from 'src/engine/core-modules/file/file-upload/guards/create-file-upload-permission.guard';
 import { FileUploadService } from 'src/engine/core-modules/file/file-upload/services/file-upload.service';
 import { PreventNestToAutoLogGraphqlErrorsFilter } from 'src/engine/core-modules/graphql/filters/prevent-nest-to-auto-log-graphql-errors.filter';
 import { ResolverValidationPipe } from 'src/engine/core-modules/graphql/pipes/resolver-validation.pipe';
@@ -32,7 +33,7 @@ export class FileUploadResolver {
   constructor(private readonly fileUploadService: FileUploadService) {}
 
   @Mutation(() => FileUploadTargetDTO)
-  @UseGuards(SettingsPermissionGuard(PermissionFlagType.UPLOAD_FILE))
+  @UseGuards(CreateFileUploadPermissionGuard)
   async createFileUpload(
     @AuthWorkspace()
     { id: workspaceId }: WorkspaceEntity,